+++ amethyst_utils/src/render.rs
@@ -1,8 +1,8 @@
 //! Module containing utility methods to easily create `Material` and `Texture` handles using a minimal amount of code.
 
-use amethyst_assets::{Loader, AssetStorage, Handle};
-use amethyst_renderer::{Texture, MaterialDefaults, Material, PngFormat, TextureMetadata};
+use amethyst_assets::{AssetStorage, Handle, Loader};
 use amethyst_core::specs::World;
+use amethyst_renderer::{Material, MaterialDefaults, PngFormat, Texture, TextureMetadata};
 
 /// Generate a `Material` from a color.
 pub fn material_from_color(
@@ -38,10 +38,20 @@ pub fn material_from_png(
 
 /// Generate a `Material` from a color.
 pub fn world_material_from_color(color: [f32; 4], world: &World) -> Material {
-	material_from_color(color, &world.read_resource(), &world.read_resource(), &world.read_resource())
+    material_from_color(
+        color,
+        &world.read_resource(),
+        &world.read_resource(),
+        &world.read_resource(),
+    )
 }
 
 /// Generate a `Material` from a path pointing to a png image.
 pub fn world_material_from_png(path: String, world: &World) -> Material {
-	material_from_png(path, &world.read_resource(), &world.read_resource(), &world.read_resource())
+    material_from_png(
+        path,
+        &world.read_resource(),
+        &world.read_resource(),
+        &world.read_resource(),
+    )
 }